+++ examples/p2p/src/model.rs
@@ -1,9 +1,13 @@
+use wavesyncdb::{CrudModel, derive::{CrudModel, SyncedCrudModel}};
+use wavesyncdb::SyncedModel;
 
+#[derive(CrudModel, SyncedCrudModel, Debug, Default)]
 pub struct Task {
-    pub id: Option<i32>,
+    #[primary_key]
+    pub id: i32,
     pub title: String,
     pub description: Option<String>,
     pub completed: bool,
-    pub created_at: Option<chrono::NaiveDateTime>,
-    pub updated_at: Option<chrono::NaiveDateTime>,
+    // pub created_at: Option<chrono::NaiveDateTime>,
+    // pub updated_at: Option<chrono::NaiveDateTime>,
 }